language/haskell: move to compat
